+++ src/components/SystemInsights.vue
@@ -2,14 +2,15 @@
 import { computed, nextTick, onMounted, onUnmounted, ref, watch } from 'vue';
 import { useRoute } from 'vue-router';
 
+import LoadingModule from './modules/LoadingModule.vue';
 import ExternalSystem from './ExternalSystem.vue';
-import { tryImportWithRetries } from '../utils/moduleFederation';
+import { tryImportWithRetries } from '@/utils/moduleFederation';
 import { useSharedStore } from '@/store/Shared';
-import { useFeatureFlagsStore } from '../store/featureFlags';
-
-import LogoWeniAnimada4 from '@/assets/LogoWeniAnimada4.svg';
+import { useFeatureFlagsStore } from '@/store/featureFlags';
+import { useModuleUpdateRoute } from '@/composables/useModuleUpdateRoute';
 
 const insightsApp = ref(null);
+const insightsRouter = ref(null);
 const useIframe = ref(false);
 const iframeInsights = ref(null);
 
@@ -26,13 +27,17 @@ const route = useRoute();
 const sharedStore = useSharedStore();
 const featureFlagsStore = useFeatureFlagsStore();
 
+const { getInitialModuleRoute } = useModuleUpdateRoute('insights');
+
 async function mount({ force = false } = {}) {
+  if (!force && !props.modelValue) {
+    return;
+  }
+
   if (!featureFlagsStore.flags.insightsMF) {
     fallbackToIframe();
     return;
   }
-
-  if (!force && !props.modelValue) return;
 
   const mountInsightsApp = await tryImportWithRetries(
     () => import('insights/main'),
@@ -44,10 +49,15 @@ async function mount({ force = false } = {}) {
     return;
   }
 
-  insightsApp.value = await mountInsightsApp({
+  const initialRoute = getInitialModuleRoute();
+
+  const { app, router } = await mountInsightsApp({
     containerId: 'insights-app',
-    routerBase: `/projects/${sharedStore.current.project.uuid}/insights`,
+    initialRoute,
   });
+
+  insightsApp.value = app;
+  insightsRouter.value = router;
 }
 
 function fallbackToIframe() {
@@ -62,11 +72,11 @@ function unmount() {
   insightsApp.value = null;
 }
 
-function remount() {
+async function remount() {
+  await insightsRouter.value.replace({ name: 'home' });
   unmount();
-  nextTick(() => {
-    mount({ force: true });
-  });
+  await nextTick();
+  mount({ force: true });
 }
 
 onMounted(() => {
@@ -76,7 +86,7 @@ onMounted(() => {
 watch(
   () => props.modelValue,
   () => {
-    if (!insightsApp.value) {
+    if (props.modelValue && !insightsApp.value) {
       mount();
     }
   },
@@ -100,18 +110,14 @@ onUnmounted(() => {
 </script>
 
 <template>
+  <LoadingModule
+    data-testid="insights-loading"
+    :isModuleRoute="isInsightsRoute"
+    :hasModuleApp="!!insightsApp"
+    :useIframe="useIframe"
+  />
+
   <template v-if="sharedStore.auth.token && sharedStore.current.project.uuid">
-    <section
-      v-if="!insightsApp && isInsightsRoute && !useIframe"
-      class="system-insights__loading"
-      data-testid="insights-loading"
-    >
-      <img
-        width="64"
-        :src="LogoWeniAnimada4"
-        data-testid="insights-loading-image"
-      />
-    </section>
     <section
       v-if="!useIframe"
       v-show="insightsApp && isInsightsRoute"
@@ -135,15 +141,6 @@ onUnmounted(() => {
 <style scoped lang="scss">
 .system-insights__system {
   height: 100%;
-}
-
-.system-insights__loading {
-  height: 100%;
-  width: 100%;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
 }
 
 .system-insights__iframe {