+++ src/main.js
@@ -106,9 +106,9 @@ app.mixin({
   },
 });
 
+app.use(createPinia());
 app.use(router);
 app.use(store);
-app.use(createPinia());
 app.use(i18n);
 app.use(UnnnicSystem);